+++ ImgReconstruction/Core/IInterpolationKernel.h
@@ -9,8 +9,8 @@ class IInterpolationKernel : public IImageSubprocessor
 {
 public:
 	virtual ~IInterpolationKernel() {};
-	virtual double operator()(double x) const = 0;
-	virtual std::vector<double> Coeffs(double shift) const = 0;
+	virtual float operator()(float x) const = 0;
+	virtual std::vector<float> Coeffs(float shift) const = 0;
 
 	virtual int A() const = 0;
 };
